sgrep_lint: refactor separate branching flows
Before this commit most work was being done in a single function
in sgrep_lint even if the commands were largely separate. In particular
the --validate and --dump-ast options did not need to go through
the entire program as they are self contained.
