+++ sgrep_lint/semgrep/constants.py
@@ -10,3 +10,5 @@ DEFAULT_CONFIG_FILE = f".{DEFAULT_SGREP_CONFIG_NAME}.yml"
 DEFAULT_CONFIG_FOLDER = f".{DEFAULT_SGREP_CONFIG_NAME}"
 
 YML_EXTENSIONS = {".yml", ".yaml"}
+
+SGREP_PATH = "sgrep"
